[refactor] 커스텀 에러와 글로벌 에러 분리 #258
